Code for testing a single left turn signal.

+++ UOSM-Core/Inc/ApplicationTypes.h
@@ -105,13 +105,13 @@ typedef struct {
 } pressure_t;
 
 typedef union {
-    uint16_t all;
+    uint32_t all;
     struct {
-        uint16_t hazards_enabled:1;
-        uint16_t left_turn_enabled:1;
-        uint16_t right_turn_enabled:1;
-        uint16_t headlights_enabled:1;
-        uint16_t low_beams_enabled:1;
+        uint32_t hazards_enabled:1;
+        uint32_t left_turn_enabled:1;
+        uint32_t right_turn_enabled:1;
+        uint32_t headlights_enabled:1;
+        uint32_t low_beams_enabled:1;
     };
 } lights_status_t;
